Split out bootstrap function into new file
